+++ .travis.yml
@@ -2,5 +2,5 @@ notifications: {email: false}
 language: python
 dist: xenial
 python: 3.7
-# before_script: pip install -r requirements.txt -r requirements.dev.txt
-script: true  # TODO
+install: pip install -r requirements.txt -r requirements.dev.txt
+script: ./bin/test.sh
