+++ bin/test.sh
@@ -1,0 +1,21 @@
+#!/usr/bin/env sh
+
+exit=0
+
+checked() {
+  echo "$ $@"
+  "$@"
+  last="$?"
+  if [ "$last" -ne 0 ]; then
+    echo "$@: exit $last"
+    exit=1
+  fi
+}
+
+cd $(dirname "$0")/..
+
+checked pytest
+checked black --check tagbot test
+checked mypy tagbot test
+
+exit $exit
