+++ requirements.dev.txt
@@ -1,0 +1,3 @@
+black
+mypy
+pytest
