+++ requirements.txt
@@ -1,0 +1,2 @@
+PyGithub==1.44
+toml==0.10
