+++ setup.cfg
@@ -1,0 +1,8 @@
+[mypy]
+python_version = 3.7
+
+[mypy-github.*]
+ignore_missing_imports = True
+
+[tool:pytest]
+filterwarnings = ignore::DeprecationWarning
